requirements.txt
aws-cdk-lib
constructs
aws_solutions_constructs.aws_s3_stepfunctions